% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMicroSigScore.R
\name{calcMicroSigScore}
\alias{calcMicroSigScore}
\title{Function to calculate microbe signature scores given a set of up and down
regulated microbes and the relative abundances of microbe per sample}
\usage{
calcMicroSigScore(taxa, sig.up, sig.down)
}
\arguments{
\item{taxa}{a data frame where columns are microbes and rows are samples, first
column named 'sample'}

\item{sig.up}{a character vector listing microbes in given signature that are
up-regulated, empty if none}

\item{sig.down}{a character vector listing microbes in given signature that are
down-regulated, empty if none}
}
\value{
a data frame with samples and corresponding microbe signature scores
}
\description{
Function to calculate microbe signature scores given a set of up and down
regulated microbes and the relative abundances of microbe per sample
}
