% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSigList.R
\name{getSigList}
\alias{getSigList}
\title{Given a single cell value from BugSigDB output in the column 'MetaPhlAn.taxon.names',
returns a list of the microbes}
\usage{
getSigList(sig)
}
\arguments{
\item{sig}{character string from BugSigDB output containing comma separated
Metaphlan taxonomic names}
}
\value{
sig.list , a character list of containing the string after the last
"|" in each comma separated Metaphlan taxonomic name
}
\description{
Given a single cell value from BugSigDB output in the column 'MetaPhlAn.taxon.names',
returns a list of the microbes
}
