% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbeEntries.R
\name{microbeEntries}
\alias{microbeEntries}
\title{Microbe Entries
A function that can be used to call of specific microbe signatures.}
\usage{
microbeEntries(query)
}
\arguments{
\item{query}{Name of the signature group, or the string "all"}
}
\value{
The microbes within a desired signature group in the format
"x__microbe name" with x indicating taxonomic level. If query is passed "all"
function will return all the signature groups
}
\description{
Microbe Entries
A function that can be used to call of specific microbe signatures.
}
\examples{
microbeEntries("paa_up")
}
